Add SchoolModal component, naming change
