--- packages/outside/src/modals/CreateModal/CreateSchoolModal/index.stories.tsx
@@ -1,13 +1,0 @@
-'use client'
-
-import CreateSchoolModal from '.'
-import type { Meta, StoryObj } from '@storybook/react'
-
-export default {
-  title: 'outside/modals/CreateSchoolModal',
-  component: CreateSchoolModal,
-} as Meta<typeof CreateSchoolModal>
-
-type Story = StoryObj<typeof CreateSchoolModal>
-
-export const Primary: Story = {}
