+++ packages/outside/src/modals/CreateModal/SchoolModal/index.stories.tsx
@@ -1,0 +1,13 @@
+'use client'
+
+import SchoolModal from '.'
+import type { Meta, StoryObj } from '@storybook/react'
+
+export default {
+  title: 'outside/modals/SchoolModal',
+  component: SchoolModal,
+} as Meta<typeof SchoolModal>
+
+type Story = StoryObj<typeof SchoolModal>
+
+export const Primary: Story = {}
