+++ packages/outside/src/modals/CreateModal/SchoolModal/style.ts
